Scatter now works with zero copy numpy arrays. Also minor corrections to scatter types test

+++ pupympi/tests/TEST_scatter_types.py
@@ -50,19 +50,18 @@ my_data = world.scatter(scatter_data, root=SCATTER_ROOT)
 assert numpy.alltrue(my_data == numpy.arange(rank*chunksize,(rank+1)*chunksize))
 
 
-# multidimensional Numpy float arrays
+# multidimensional Numpy int arrays
 chunksize = 3
 if rank == SCATTER_ROOT:
-    #2*3*2 ints of 8 bytes = 96 bytes = 48 bytes per proc
     scatter_data = numpy.arange(size*chunksize*2*5).reshape(size*2,chunksize,5)
 else:
     scatter_data = None
 my_data = world.scatter(scatter_data, root=SCATTER_ROOT)
-assert numpy.alltrue(my_data == numpy.arange(rank*chunksize*2*5,(rank+1)*chunksize*2*5).reshape(2,chunksize,5) )
+assert numpy.alltrue(my_data == numpy.arange(rank*chunksize*2*5,(rank+1)*chunksize*2*5,dtype=numpy.float).reshape(2,chunksize,5) )
 
 # 4D Numpy float array
 if rank == SCATTER_ROOT:    
-    scatter_data = numpy.arange(size*chunksize*12).reshape(size*3,chunksize,2,2)
+    scatter_data = numpy.arange(size*chunksize*12,dtype=numpy.float).reshape(size*3,chunksize,2,2)
 else:
     scatter_data = None
